♻️ 統計情報を表示するStatsSummaryコンポーネントを追加
